adding stream to its own frame class for readability
